pyproject.toml
[tool.poetry]
name = "e-affidavit"
version = "0.1.0"
description = ""
authors = ["AlexAdedeji <alex.adedeji@identikosolutions.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
alembic = "^1.13.1"
fastapi = {extras = ["all"], version = "^0.109.2"}
loguru = "^0.7.2"
pydantic-settings = "^2.1.0"
psycopg2 = "^2.9.9"
motor = "^3.3.2"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
postmarker = "^1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
